backend/requirements.txt
flask
flask-cors
torch
torchvision
timm
pillow
sqlalchemy
pyjwt
werkzeug
google-generativeai
python-dotenv
